Buy Loeje

+++ src/main/java/com/gmail/markushygedombrowski/npc/vagthavende/DeliverGearGUI.java
@@ -70,7 +70,7 @@ public class DeliverGearGUI implements Listener {
     private boolean gear(Player p, ItemStack clickeditem) {
         if (deliverGear(p, clickeditem)) {
             PlayerProfile profile = playerProfiles.getPlayerProfile(p.getUniqueId());
-            profile.setXp((int)profile.getProperty("exp") + itemProfileLoader.getItemProfile(clickeditem.getItemMeta().getDisplayName()).getExp());
+            profile.setXp(((Double) profile.getProperty("exp")).intValue() + itemProfileLoader.getItemProfile(clickeditem.getItemMeta().getDisplayName()).getExp());
 
             return true;
         }
